# src/modules/projects/routes.ts
import {Router} from 'express'; import {z} from 'zod'; import {prisma} from '../../config/database'; import {auth,admin} from '../../middleware/auth'; import {AppError} from '../../utils/errors';
const r=Router(); const body=z.object({name:z.string().min(1).max(120),description:z.string().max(5000).optional()});
r.use(auth);
r.post('/',async(req,res,next)=>{try{const b=body.parse(req.body);res.status(201).json(await prisma.project.create({data:{...b,organizationId:req.auth!.organizationId}}))}catch(e){next(e)}});
r.get('/',async(req,res,next)=>{try{const page=Math.max(1,Number(req.query.page??1)),limit=Math.min(100,Math.max(1,Number(req.query.limit??20)));const where={organizationId:req.auth!.organizationId,deletedAt:null};const [data,total]=await prisma.$transaction([prisma.project.findMany({where,skip:(page-1)*limit,take:limit,orderBy:{createdAt:'desc'}}),prisma.project.count({where})]);res.json({data,total,page,limit})}catch(e){next(e)}});
r.get('/:id/dashboard',async(req,res,next)=>{try{const p=await prisma.project.findFirst({where:{id:String(req.params.id),organizationId:req.auth!.organizationId,deletedAt:null}});if(!p)throw new AppError(403,'PROJECT_FORBIDDEN','Forbidden');const counts=await prisma.task.groupBy({by:['status'],where:{projectId:p.id,deletedAt:null},_count:true});res.json({projectId:p.id,tasks:Object.fromEntries(['todo','in_progress','review','done'].map(s=>[s,counts.find(x=>x.status===s)?._count??0]))})}catch(e){next(e)}});
r.get('/:id',async(req,res,next)=>{try{const p=await prisma.project.findFirst({where:{id:String(req.params.id),organizationId:req.auth!.organizationId,deletedAt:null},include:{_count:{select:{tasks:true}}}});if(!p)throw new AppError(403,'PROJECT_FORBIDDEN','Forbidden');res.json(p)}catch(e){next(e)}});
r.patch('/:id',async(req,res,next)=>{try{const b=body.partial().parse(req.body);const p=await prisma.project.findFirst({where:{id:String(req.params.id),organizationId:req.auth!.organizationId,deletedAt:null}});if(!p)throw new AppError(403,'PROJECT_FORBIDDEN','Forbidden');res.json(await prisma.project.update({where:{id:p.id},data:b}))}catch(e){next(e)}});
r.delete('/:id',admin,async(req,res,next)=>{try{const p=await prisma.project.findFirst({where:{id:String(req.params.id),organizationId:req.auth!.organizationId,deletedAt:null}});if(!p)throw new AppError(403,'PROJECT_FORBIDDEN','Forbidden');await prisma.project.update({where:{id:p.id},data:{deletedAt:new Date()}});res.status(204).send()}catch(e){next(e)}}); export default r;